refeactoring creport, more stream, less FILE make cr_output_region testable

+++ src/util/crmessage.h
@@ -21,7 +21,7 @@ extern "C" {
     int cr_ignore(variant v, const char *name, char *buffer, const void *userdata);
 
     void crt_register(const struct message_type *mtype);
-    int cr_render(const struct message *msg, char *buffer,
+    size_t cr_render(const struct message *msg, char *buffer,
         const void *userdata);
 
 #ifdef __cplusplus
